// frontend/src/shared/api/generated/model/lessonBlockSchema.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */

export interface LessonBlockSchema {
  id: string;
  lesson_id: string;
  user_id: string | null;
  content: string;
  block_number: number;
  block_id?: string | null;
  title?: string | null;
  is_summary: boolean | null;
  progress?: number;
  created_at: string;
}